+++ src/BearerAuthenticator.php
@@ -5,6 +5,8 @@ use Psr\Http\Message\ServerRequestInterface as PsrServerRequestInterface;
 use Pyncer\Access\AbstractBearerAuthenticator;
 use Pyncer\Data\Mapper\MapperAdaptorInterface;
 
+use const Pyncer\DATE_TIME_NOW as PYNCER_DATE_TIME_NOW;
+
 class BearerAuthenticator extends AbstractBearerAuthenticator
 {
     public function __construct(
@@ -12,11 +14,31 @@ class BearerAuthenticator extends AbstractBearerAuthenticator
         protected MapperAdaptorInterface $userMapperAdaptor,
         PsrServerRequestInterface $request,
         string $realm,
+        protected ?int $guestUserId = null,
     ) {
         parent::__construct($request, $realm);
     }
 
+    public function getGuestUserId(): ?int
+    {
+        return $this->guestUserId;
+    }
+
     public function authenticate(string $token): bool
+    {
+        if (!$this->authenticateToken($token)) {
+            if ($this->guestUserId) {
+                $this->guestUserModel = $this->userMapperAdaptor->getMapper()->selectById(
+                    $this->guestUserId,
+                    $this->userMapperAdaptor->getMapperQuery()
+                );
+            }
+        }
+
+        return true;
+    }
+
+    protected function authenticateToken(string $token): bool
     {
         $tokenModel = $this->tokenMapperAdaptor->getMapper()->selectByColumns(
             $this->tokenMapperAdaptor->getFormatter()->formatData([
@@ -37,7 +59,13 @@ class BearerAuthenticator extends AbstractBearerAuthenticator
             $tokenModel->getData()
         );
 
-        if (!$token['user_id'] ?? null) {
+        if ($token['expiration_date_time'] ?? null) {
+            if ($token['expiration_date_time'] < PYNCER_DATE_TIME_NOW) {
+                return false;
+            }
+        }
+
+        if (!($token['user_id'] ?? null)) {
             return true;
         }
 
@@ -54,4 +82,5 @@ class BearerAuthenticator extends AbstractBearerAuthenticator
 
         return true;
     }
+
 }