+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.1.0 - 2023-06-14
+
+### Added
+
+- Added ability to sepcify a guest user model.
+- PHPStan static analysis.
+
+### Fixed
+
+- Fixed bearer authenticator not taking into consideration token expiration date.
+- Miscellaneous issues brought up by PHPStan.
+
 ## 1.0.0 - 2022-12-28
 
 Initial release.